lint: split big file
