+++ lib/services/media/media_pod_paths.dart
@@ -1,6 +1,6 @@
 /// Path constants for media (audio / video) storage in the Solid Pod.
 ///
-// Time-stamp: <2026-02-28 GitHub Copilot>
+// Time-stamp: <2026-02-28 Miduo>
 ///
 /// Copyright (C) 2026, Software Innovation Institute, ANU.
 ///
