+++ lib/services/media/media_pod_service.dart
@@ -1,6 +1,6 @@
 /// Solid Pod service for managing audio and video media items.
 ///
-// Time-stamp: <2026-02-28 GitHub Copilot>
+// Time-stamp: <2026-02-28 Miduo>
 ///
 /// Copyright (C) 2026, Software Innovation Institute, ANU.
 ///
