+++ lib/utils/fullscreen_stub.dart
@@ -1,6 +1,6 @@
 /// System fullscreen helpers — stub for NATIVE (non-web) platforms.
 ///
-// Time-stamp: <2026-02-28 GitHub Copilot>
+// Time-stamp: <2026-02-28 Miduo>
 ///
 /// Copyright (C) 2026, Software Innovation Institute, ANU.
 ///
