+++ lib/widgets/media/upload_media_dialog.dart
@@ -1,6 +1,6 @@
 /// Dialog for uploading an audio or video file to the user's Solid Pod.
 ///
-// Time-stamp: <2026-02-28 GitHub Copilot>
+// Time-stamp: <2026-02-28 Miduo>
 ///
 /// Copyright (C) 2026, Software Innovation Institute, ANU.
 ///
